+++ tests/ui/widgets/test_input.py
@@ -2,20 +2,11 @@ import pytest
 
 pytest.importorskip('playwright')
 
-import datetime as dt
-
 from panel.tests.util import serve_component, wait_until
-from panel_material_ui.widgets import (
-    DatePicker,
-    TextInput, PasswordInput, TextAreaInput, Checkbox, Switch, FileInput, ToggleIcon,
-)
+from panel_material_ui.widgets import TextInput, PasswordInput
 from playwright.sync_api import expect
 
 pytestmark = pytest.mark.ui
-
-
-# observe when serving the component
-TEXTAREA_LINE_HEIGHT = 23
 
 
 @pytest.mark.parametrize('variant', ["filled", "outlined", "standard"])
@@ -27,7 +18,6 @@ def test_text_input_variant(page, variant):
         expect(page.locator('.MuiInput-root')).to_have_count(1)
     else:
         expect(page.locator(f'.Mui{variant.capitalize()}Input-root')).to_have_count(1)
-
 
 def test_text_input_typing(page):
     widget = TextInput(name='Test', placeholder='Type something...')
@@ -61,7 +51,6 @@ def test_text_input_typing(page):
     input_field.press('Enter')
     wait_until(lambda: widget.value == 'Hello World Again', page)
 
-
 @pytest.mark.from_panel
 def test_textinput_enter_pressed(page):
     text_input = TextInput()
@@ -81,7 +70,6 @@ def test_textinput_enter_pressed(page):
     input_area.press("Enter")
     wait_until(lambda: clicks[0] == 2)
 
-
 def test_textinput_max_length(page):
     widget = TextInput(max_length=2)
     serve_component(page, widget)
@@ -94,7 +82,6 @@ def test_textinput_max_length(page):
     expect(input_area).to_have_value("12")
     wait_until(lambda: widget.value_input == "12", page)
 
-
 def test_password_show_hide(page):
     widget = PasswordInput(label='Password', placeholder='Enter your password here ...')
     serve_component(page, widget)
@@ -105,7 +92,6 @@ def test_password_show_hide(page):
     eye_button.click()
     # password is displayed
     expect(page.locator('input[type="text"]')).to_have_count(1)
-
 
 def test_password_max_length(page):
     widget = PasswordInput(max_length=2)
@@ -118,313 +104,3 @@ def test_password_max_length(page):
     input_area.type("123")
     expect(input_area).to_have_value("12")
     wait_until(lambda: widget.value_input == "12", page)
-
-
-def test_text_area_input(page):
-    widget = TextAreaInput(label='Description', placeholder='Enter your description here...', rows=5)
-    serve_component(page, widget)
-    expect(page.locator('.text-area-input')).to_have_count(1)
-    expect(page.locator('textarea[rows="5"]')).to_have_count(1)
-
-
-def test_text_area_typing(page):
-    widget = TextAreaInput(label='Description', placeholder='Type something...')
-    serve_component(page, widget)
-
-    # Find the textarea and type into it
-    textarea = page.locator('textarea').nth(0)
-    textarea.click()
-
-    # Type text including newlines
-    textarea.type('Multiline', delay=50)
-    textarea.press('Enter')
-    textarea.type('Text', delay=50)
-    textarea.press('Enter')
-    textarea.type('Test', delay=50)
-
-    # Check that the text appears as we type (value_input)
-    expect(textarea).to_have_value('Multiline\nText\nTest')
-
-    # Check value_input is updated while typing
-    wait_until(lambda: widget.value_input == 'Multiline\nText\nTest', page)
-
-    # But value should still be the original value (empty string) since we haven't blurred
-    assert widget.value == ''
-
-    # Click elsewhere to trigger blur
-    page.locator('body').click()
-    wait_until(lambda: widget.value == 'Multiline\nText\nTest', page)
-
-
-@pytest.mark.from_panel
-def test_text_area_auto_grow(page):
-    widget = TextAreaInput(auto_grow=True, value="1\n2\n3\n4\n")
-    serve_component(page, widget)
-
-    input_area = page.locator('.MuiInputBase-input').nth(0)
-    input_area.click()
-    input_area.press('Enter')
-    input_area.press('Enter')
-    input_area.press('Enter')
-
-    # 8 rows
-    wait_until(lambda: input_area.bounding_box()['height'] == 8 * TEXTAREA_LINE_HEIGHT, page)
-
-
-@pytest.mark.from_panel
-def test_text_area_auto_grow_max_rows(page):
-    text_area = TextAreaInput(auto_grow=True, value="1\n2\n3\n4\n", max_rows=7)
-
-    serve_component(page, text_area)
-
-    input_area = page.locator('.MuiInputBase-input').nth(0)
-    input_area.click()
-    input_area.press('Enter')
-    input_area.press('Enter')
-    input_area.press('Enter')
-
-    wait_until(lambda: input_area.bounding_box()['height'] == 7 * TEXTAREA_LINE_HEIGHT, page)
-
-
-@pytest.mark.from_panel
-def test_text_area_auto_grow_min_rows(page):
-    text_area = TextAreaInput(auto_grow=True, value="1\n2\n3\n4\n", rows=3)
-    serve_component(page, text_area)
-
-    input_area = page.locator('.MuiInputBase-input').nth(0)
-    input_area.click()
-    for _ in range(5):
-        input_area.press('ArrowDown')
-    for _ in range(10):
-        input_area.press('Backspace')
-
-    wait_until(lambda: input_area.bounding_box()['height'] == 3 * TEXTAREA_LINE_HEIGHT, page)
-
-
-@pytest.mark.from_panel
-def test_text_area_auto_grow_shrink_back_on_new_value(page):
-    text_area = TextAreaInput(auto_grow=True, value="1\n2\n3\n4\n", max_rows=5)
-    serve_component(page, text_area)
-
-    input_area = page.locator('.MuiInputBase-input').nth(0)
-    input_area.click()
-    for _ in range(5):
-        input_area.press('ArrowDown')
-    for _ in range(10):
-        input_area.press('Backspace')
-
-    text_area.value = ""
-    assert input_area.bounding_box()['height'] == 2 * TEXTAREA_LINE_HEIGHT
-
-
-def test_text_area_max_length(page):
-    widget = TextAreaInput(max_length=2)
-    serve_component(page, widget)
-
-    # Find the input field and type into it
-    input_area = page.locator('.MuiInputBase-input').nth(0)
-    input_area.click()
-    # type more but only first max_length characters are allowed
-    input_area.type("123")
-    expect(input_area).to_have_value("12")
-    wait_until(lambda: widget.value_input == "12", page)
-
-
-def test_checkbox(page):
-    widget = Checkbox(label='Works with the tools you know and love', value=True)
-    serve_component(page, widget)
-    expect(page.locator('.checkbox')).to_have_count(1)
-
-
-def test_checkbox_basic_functionality(page):
-    widget = Checkbox(label='Test Checkbox', value=False)
-    serve_component(page, widget)
-
-    # Initial state
-    expect(page.locator('.checkbox')).to_have_count(1)
-    expect(page.locator('.MuiCheckbox-root')).to_have_count(1)
-    expect(page.locator('.MuiFormControlLabel-root')).to_have_text('Test Checkbox')
-    expect(page.locator('.MuiCheckbox-root')).not_to_be_checked()
-
-    # Click to check
-    page.locator('.MuiCheckbox-root').click()
-    wait_until(lambda: widget.value is True, page)
-    expect(page.locator('.MuiCheckbox-root')).to_be_checked()
-
-    # Click to uncheck
-    page.locator('.MuiCheckbox-root').click()
-    wait_until(lambda: widget.value is False, page)
-    expect(page.locator('.MuiCheckbox-root')).not_to_be_checked()
-
-
-@pytest.mark.parametrize('color', ["primary", "secondary", "error", "info", "success", "warning"])
-def test_checkbox_colors(page, color):
-    widget = Checkbox(label='Test Checkbox', color=color)
-    serve_component(page, widget)
-
-    expect(page.locator('.checkbox')).to_have_count(1)
-    expect(page.locator(f'.MuiCheckbox-color{color.capitalize()}')).to_have_count(1)
-
-
-@pytest.mark.parametrize('size', ["small", "medium", "large"])
-def test_checkbox_sizes(page, size):
-    widget = Checkbox(label='Test Checkbox', size=size)
-    serve_component(page, widget)
-
-    expect(page.locator('.checkbox')).to_have_count(1)
-    expect(page.locator(f'.MuiCheckbox-size{size.capitalize()}')).to_have_count(1)
-
-
-def test_checkbox_disabled_state(page):
-    widget = Checkbox(label='Test Checkbox', disabled=True)
-    serve_component(page, widget)
-
-    expect(page.locator('.checkbox')).to_have_count(1)
-    expect(page.locator('.MuiCheckbox-root')).to_be_disabled()
-
-
-def test_checkbox_indeterminate_state(page):
-    widget = Checkbox(label='Test Checkbox', indeterminate=True, value=None)
-    serve_component(page, widget)
-
-    expect(page.locator('.checkbox')).to_have_count(1)
-    expect(page.locator('.PrivateSwitchBase-input')).to_have_attribute('data-indeterminate', 'true')
-
-    # Click should change to checked state
-    page.locator('.MuiCheckbox-root').click()
-    wait_until(lambda: widget.value is True, page)
-    expect(page.locator('.MuiCheckbox-root')).to_be_checked()
-    expect(page.locator('.PrivateSwitchBase-input')).not_to_have_attribute('data-indeterminate', 'true')
-
-
-def test_checkbox_value_updates(page):
-    widget = Checkbox(label='Test Checkbox', value=False)
-    serve_component(page, widget)
-
-    # Initial state
-    expect(page.locator('.MuiCheckbox-root')).not_to_be_checked()
-
-    # Update value programmatically
-    widget.value = True
-    wait_until(lambda: page.locator('.MuiCheckbox-root').is_checked(), page)
-
-    # Update to indeterminate
-    widget.indeterminate = True
-    widget.value = None
-    wait_until(lambda: page.locator('.PrivateSwitchBase-input').get_attribute('data-indeterminate') == 'true', page)
-
-
-def test_checkbox_label_click(page):
-    widget = Checkbox(label='Clickable Label', value=False)
-    serve_component(page, widget)
-
-    # Initial state
-    expect(page.locator('.MuiCheckbox-root')).not_to_be_checked()
-
-    # Click label to check
-    page.locator('.MuiFormControlLabel-root').click()
-    wait_until(lambda: widget.value is True, page)
-    expect(page.locator('.MuiCheckbox-root')).to_be_checked()
-
-    # Click label to uncheck
-    page.locator('.MuiFormControlLabel-root').click()
-    wait_until(lambda: widget.value is False, page)
-    expect(page.locator('.MuiCheckbox-root')).not_to_be_checked()
-
-
-def test_checkbox_no_label(page):
-    widget = Checkbox(value=False)
-    serve_component(page, widget)
-
-    expect(page.locator('.checkbox')).to_have_count(1)
-    expect(page.locator('.MuiFormControlLabel-root')).to_have_text('')
-    expect(page.locator('.MuiCheckbox-root')).to_have_count(1)
-
-    # Verify checkbox still works without label
-    page.locator('.MuiCheckbox-root').click()
-    wait_until(lambda: widget.value is True, page)
-    expect(page.locator('.MuiCheckbox-root')).to_be_checked()
-
-
-def test_switch(page):
-    widget = Switch(label='Works with the tools you know and love', value=True)
-    serve_component(page, widget)
-    expect(page.locator('.switch')).to_have_count(1)
-
-
-def test_fileinput(page):
-    widget = FileInput(accept='.png,.jpeg', multiple=True)
-    serve_component(page, widget)
-    expect(page.locator('.file-input')).to_have_count(1)
-
-
-def test_toggle_icon(page):
-    widget = ToggleIcon(icon="thumb-up", active_icon="thumb-down", size="small", description="Like")
-    serve_component(page, widget)
-
-    expect(page.locator('.toggle-icon')).to_have_count(1)
-    icon = page.locator('.MuiCheckbox-root')
-    expect(icon).to_have_text("thumb-up")
-    icon.click()
-    expect(icon).to_have_text("thumb-down")
-
-
-def test_datepicker_enabled_dates(page):
-    widget = DatePicker(
-        name='Date Picker',
-        start=dt.date(2024, 4, 1),
-        end=dt.date(2024, 4, 20),
-        enabled_dates=[
-            dt.date(2024, 4, 1),
-            dt.date(2024, 4, 2),
-            dt.date(2024, 4, 20),
-            dt.date(2024, 4, 21),
-        ]
-    )
-    serve_component(page, widget)
-
-    # click the datepicker icon to show dates to select
-    icon = page.locator(".MuiIconButton-root")
-    icon.click()
-    # Select all buttons in the date picker
-    buttons = page.locator('.MuiPickersDay-root').all()
-
-    enabled_buttons = []
-    for button in buttons:
-        is_disabled = button.get_attribute("disabled") is not None or button.evaluate(
-            "el => el.classList.contains('Mui-disabled')")
-        if not is_disabled:
-            value = button.inner_text()  # Get the button's displayed text (date)
-            enabled_buttons.append(value)
-
-    # only enabled dates within the start to end range are selectable
-    assert enabled_buttons == ['1', '2', '20']
-
-
-def test_datepicker_disabled_dates(page):
-    widget = DatePicker(
-        name='Date Picker',
-        start=dt.date(2024, 4, 1),
-        end=dt.date(2024, 4, 20),
-        disabled_dates=[
-            dt.date(2024, 4, i) for i in range(1, 18)
-        ]
-    )
-    serve_component(page, widget)
-
-    # click the datepicker icon to show dates to select
-    icon = page.locator(".MuiIconButton-root")
-    icon.click()
-    # Select all buttons in the date picker
-    buttons = page.locator('.MuiPickersDay-root').all()
-
-    enabled_buttons = []
-    for button in buttons:
-        is_disabled = button.get_attribute("disabled") is not None or button.evaluate(
-            "el => el.classList.contains('Mui-disabled')")
-        if not is_disabled:
-            value = button.inner_text()  # Get the button's displayed text (date)
-            enabled_buttons.append(value)
-
-    # only enabled dates within the start to end range are selectable
-    assert enabled_buttons == ['18', '19', '20']
